Strawman: avoid `String` allocations
This is a strawman proposal on top of the previous commit to show a
plausible reshuffle that avoids the String allocation (1 per cell).

The main constraint is that something has to own the backing storage
of an `&str` (the `csv::Reader`), therefore the parsing and the
"copying into Ruby's heap" must happen inside the same stack frame
(i.e. we cannot just return `Vec<Vec<&'a str>>` as there is no valid
`'a` that would live long enough).

This is just the smallest possible delta that does the trick. It is
somewhat inelegant for serveral reasons (e.g. it breaks the previous
abstraction boundaries, leaks Ruby semantics into the Rust code  and
does not handle the possible `ToRuby` failure). There are more clever
ways to structure it to avoid some of the downsides, but this is just
to show that it _can_ be done (without writing unsafe coercion code).

More importantly, this shows the performance impact of the `String`
allocations. Surprisingly, it appears that this change is sufficient
to recoup all of the lost performance from the previous commit.

Thinking about it, it makes sense. With the right `size_hint()`,
allocating and filling the `Vec`s are quite cheap and can be easily
optimized. More importantly that operation only occurs once per row.
The `String` allocations are much more costly (since it cannot be
pre-sized) and has to happen once per cell. Therefore, it makes sense
that eliminating the `String` allocations would make a huge impact.

===================== MASTER ====================

Warming up --------------------------------------
            Ruby CSV    10.000  i/100ms
               rscsv    44.000  i/100ms
Calculating -------------------------------------
            Ruby CSV    106.381  (± 3.8%) i/s -    540.000  in   5.083809s
               rscsv    445.078  (± 3.8%) i/s -      2.244k in   5.050089s

Comparison:
               rscsv:      445.1 i/s
            Ruby CSV:      106.4 i/s - 4.18x  slower

Warming up --------------------------------------
            Ruby CSV     8.000  i/100ms
               rscsv    26.000  i/100ms
Calculating -------------------------------------
            Ruby CSV     86.912  (± 4.6%) i/s -    440.000  in   5.074026s
               rscsv    264.238  (± 4.5%) i/s -      1.326k in   5.029925s

Comparison:
               rscsv:      264.2 i/s
            Ruby CSV:       86.9 i/s - 3.04x  slower

===================== AFTER =====================

Warming up --------------------------------------
            Ruby CSV    10.000  i/100ms
               rscsv    40.000  i/100ms
Calculating -------------------------------------
            Ruby CSV    107.672  (± 2.8%) i/s -    540.000  in   5.019673s
               rscsv    460.184  (± 5.9%) i/s -      2.320k in   5.064426s

Comparison:
               rscsv:      460.2 i/s
            Ruby CSV:      107.7 i/s - 4.27x  slower

Warming up --------------------------------------
            Ruby CSV     8.000  i/100ms
               rscsv    26.000  i/100ms
Calculating -------------------------------------
            Ruby CSV     88.117  (± 3.4%) i/s -    440.000  in   5.000851s
               rscsv    268.792  (± 4.1%) i/s -      1.352k in   5.039892s

Comparison:
               rscsv:      268.8 i/s
            Ruby CSV:       88.1 i/s - 3.05x  slower